Optimize integer types and use CMSIS bit-field access
- Replace uint_fast8_t/int_fast8_t with uint8_t/int8_t for storage
- Use int32_t for hot-path loop counters to avoid extend instructions
- Change function parameters to uint8_t where values are small
- Use .bit.FIELD = 1 pattern instead of .reg |= MASK for single bits
- Add explicit casts to silence truncation warnings
- Mask PMUX values to 4 bits for bit-field assignment

Saves RAM by using appropriate sized types while maintaining
performance in DSP hot-path functions.

+++ src/driver_PORT.h
@@ -5,8 +5,8 @@
 
 /* Types */
 typedef struct Pin_ {
-  uint16_t grp;
-  uint16_t pin;
+  uint8_t grp;
+  uint8_t pin;
 } Pin_t;
 
 typedef enum PINDIR_ { PIN_DIR_IN, PIN_DIR_OUT } PINDIR_t;
@@ -23,7 +23,7 @@ typedef enum PINDRV_ { PIN_DRV_CLR, PIN_DRV_SET, PIN_DRV_TGL } PINDRV_t;
  *  @param [in] cfg : configuration option
  *  @param [in] cs  : clear or set configuration
  */
-void portPinCfg(const uint32_t grp, const uint32_t pin, const uint32_t cfg,
+void portPinCfg(const uint8_t grp, const uint8_t pin, const uint8_t cfg,
                 const PINCFG_t cs);
 
 /*! @brief Sets a pin as input or output
@@ -31,33 +31,33 @@ void portPinCfg(const uint32_t grp, const uint32_t pin, const uint32_t cfg,
  *  @param [in] pin : PIN number
  *  @param [in] mode: PIN_DIR_IN for input, PIN_DIR_OUT for output
  */
-void portPinDir(const uint32_t grp, const uint32_t pin, const PINDIR_t mode);
+void portPinDir(const uint8_t grp, const uint8_t pin, const PINDIR_t mode);
 
 /*! @brief Sets the pin driver value
  *  @param [in] grp : Group number
  *  @param [in] pin : Pin number
  *  @param [in] drv : Clear, set, or toggle pin
  */
-void portPinDrv(const uint32_t grp, const uint32_t pin, const PINDRV_t drv);
+void portPinDrv(const uint8_t grp, const uint8_t pin, const PINDRV_t drv);
 
 /*! @brief Sets the mux for pin alternate function
  *  @param [in] grp : Group number
  *  @param [in] pin : Pin number
  *  @param [in] mux : Mux mode
  */
-void portPinMux(const uint32_t grp, const uint32_t pin, const uint32_t mux);
+void portPinMux(const uint8_t grp, const uint8_t pin, const uint8_t mux);
 
 /*! @brief Clear the mux for pin alternate function
  *  @param [in] grp : Group number
  *  @param [in] pin : Pin number
  */
-void portPinMuxClear(const uint32_t grp, const uint32_t pin);
+void portPinMuxClear(const uint8_t grp, const uint8_t pin);
 
 /*! @brief Returns the pin value
  *  @param [in] grp : Group number
  *  @param [in] pin : Pin number
  */
-bool portPinValue(const uint32_t grp, const uint32_t pin);
+bool portPinValue(const uint8_t grp, const uint8_t pin);
 
 /*! @brief   Configure the ports.
  *           Ports for peripherals are configured in their setup functions
