+++ src/temperature.c
@@ -57,7 +57,7 @@ TempDev_t tempReadSerial(const TEMP_INTF_t intf, const uint8_t dev) {
   return (TempDev_t){.id = 0, .intf = TEMP_INTF_NONE};
 }
 
-TempStatus_t tempStartSample(const TEMP_INTF_t intf, const uint32_t dev) {
+TempStatus_t tempStartSample(const TEMP_INTF_t intf, const uint8_t dev) {
 
   if (0 == numSensors) {
     return TEMP_NO_SENSORS;
